Refactor StubConnection
